DB Remote Crash bug corrected? - (NSInteger)read:(uint8_t *)buffer maxLength:(NSUInteger)len can return -1 : you cannot store result in unsigned int...
git-svn-id: https://svn.code.sf.net/p/osirix/code/branches/aless_coredata@8914 4e9e6e3d-4551-47eb-9dbe-a6127f9400a1

+++ nitrogen/Sources/N2Connection.mm
@@ -304,7 +304,7 @@ NSString* N2ConnectionStatusDidChangeNotification = @"N2ConnectionStatusDidChang
 			if (_maximumReadSizePerEvent && maxLength > _maximumReadSizePerEvent-readSizeForThisEvent)
 				maxLength = _maximumReadSizePerEvent-readSizeForThisEvent;
 			uint8_t buffer[maxLength];
-			unsigned int length = [_inputStream read:buffer maxLength:maxLength];
+			NSInteger length = [_inputStream read:buffer maxLength:maxLength];
 			
 			if (length > 0) {
 				// DLog(@"%@ Read %d Bytes", self, length);
